Extract DiaryEntryKey to a separate file
